feat: add MasterStopPropagandaResolver with spdnsclient unix-like DNS client but without /etc/resolv.conf Just pure hardcoded config and []string
